Split mob drop recipe
